test(activesupport): move TS-only Instrumenter tests to instrumenter.trails.test.ts
Rails' test_record_with_exception covers Event#record, which trails' Event
does not implement — reusing that name for an instrument() exception test
mis-maps in test:compare. Async instrument has no Rails counterpart either.
Both move to the trails-only file; note the Event payload-by-reference
coupling that keeps subscribers seeing the exception keys.
